Вынес логику работы со списком пользователей коливинга в одно место
